requirements.txt
aiogram==3.0.0b7
sqlalchemy==2.0.25
python-dotenv==1.0.0
aiosqlite==0.19.0